+++ mod-source-record-manager-server/src/main/java/org/folio/dao/IncomingRecordDaoImpl.java
@@ -14,6 +14,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
+import java.util.Optional;
 import java.util.UUID;
 
 import static java.lang.String.format;
@@ -24,14 +25,30 @@ public class IncomingRecordDaoImpl implements IncomingRecordDao {
 
   private static final Logger LOGGER = LogManager.getLogger();
   public static final String INCOMING_RECORDS_TABLE = "incoming_records";
+  private static final String GET_BY_ID_SQL = "SELECT * FROM %s.%s WHERE id = $1";
   private static final String INSERT_SQL = "INSERT INTO %s.%s (id, job_execution_id, incoming_record) VALUES ($1, $2, $3)";
 
   @Autowired
   private PostgresClientFactory pgClientFactory;
 
   @Override
+  public Future<Optional<IncomingRecord>> getById(String id, String tenantId) {
+    LOGGER.debug("getById:: Get IncomingRecord by id = {} from the {} table", id, INCOMING_RECORDS_TABLE);
+    Promise<RowSet<Row>> promise = Promise.promise();
+    try {
+      String query = format(GET_BY_ID_SQL, convertToPsqlStandard(tenantId), INCOMING_RECORDS_TABLE);
+      pgClientFactory.createInstance(tenantId).selectRead(query, Tuple.of(UUID.fromString(id)), promise);
+    } catch (Exception e) {
+      LOGGER.warn("getById:: Error getting IncomingRecord by id", e);
+      promise.fail(e);
+    }
+    return promise.future().map(rowSet -> rowSet.rowCount() == 0 ? Optional.empty()
+      : Optional.of(mapRowToIncomingRecord(rowSet.iterator().next())));
+  }
+
+  @Override
   public Future<List<RowSet<Row>>> saveBatch(List<IncomingRecord> incomingRecords, String tenantId) {
-    LOGGER.info("saveBatch:: Save IncomingRecord entity to the {} table", INCOMING_RECORDS_TABLE);
+    LOGGER.debug("saveBatch:: Save IncomingRecord entity to the {} table", INCOMING_RECORDS_TABLE);
     Promise<List<RowSet<Row>>> promise = Promise.promise();
     try {
       String query = format(INSERT_SQL, convertToPsqlStandard(tenantId), INCOMING_RECORDS_TABLE);
@@ -43,6 +60,16 @@ public class IncomingRecordDaoImpl implements IncomingRecordDao {
       promise.fail(e);
     }
     return promise.future().onFailure(e -> LOGGER.warn("saveBatch:: Error saving JournalRecord entity", e));
+  }
+
+  private IncomingRecord mapRowToIncomingRecord(Row row) {
+    JsonObject jsonObject = row.getJsonObject("incoming_record");
+    return new IncomingRecord().withId(String.valueOf(row.getUUID("id")))
+      .withJobExecutionId(String.valueOf(row.getUUID("job_execution_id")))
+      .withRecordType(IncomingRecord.RecordType.fromValue(jsonObject.getString("recordType")))
+      .withOrder(jsonObject.getInteger("order"))
+      .withRawRecordContent(jsonObject.getString("rawRecordContent"))
+      .withParsedRecordContent(jsonObject.getString("parsedRecordContent"));
   }
 
   private Tuple prepareInsertQueryParameters(IncomingRecord incomingRecord) {